SAMP Results

SAMP.SAMPResult — Type

The parent of all SAMP results

SAMP.SAMPSuccess — Type

SAMPSuccess <:SAMPResult

A successful SAMP result

Field

value: the return value provided

SAMP.SAMPError — Type

SAMPError <:SAMPResult

A faulty SAMP result.

Fields

error: the error raised, as a string

user: an optional additional user message

debug: an optional longer string which may contain more detail on what went wrong. This is typically intended for debugging purposes, and may for instance be a stack trace.

code: an optional string containing a numeric or textual code identifying the error, perhaps intended to be parsable by software.

SAMP.SAMPWarning — Type

SAMPWarning <:SAMPResult

A partially successful SAMP result.

Fields

value: the value returned

warning: a structure of type SAMPError, with details on what went wrong